+++ src/storage/src/redis.h
@@ -121,11 +121,11 @@ class Redis {
   virtual Status SetsPKPatternMatchDel(const std::string& pattern, int32_t* ret);
 
   // Keys Commands
-  virtual Status StringsExpire(const Slice& key, int32_t ttl);
-  virtual Status HashesExpire(const Slice& key, int32_t ttl);
-  virtual Status ListsExpire(const Slice& key, int32_t ttl);
-  virtual Status ZsetsExpire(const Slice& key, int32_t ttl);
-  virtual Status SetsExpire(const Slice& key, int32_t ttl);
+  virtual Status StringsExpire(const Slice& key, uint64_t ttl);
+  virtual Status HashesExpire(const Slice& key, uint64_t ttl);
+  virtual Status ListsExpire(const Slice& key, uint64_t ttl);
+  virtual Status ZsetsExpire(const Slice& key, uint64_t ttl);
+  virtual Status SetsExpire(const Slice& key, uint64_t ttl);
 
   virtual Status StringsDel(const Slice& key);
   virtual Status HashesDel(const Slice& key);
@@ -168,11 +168,11 @@ class Redis {
   Status MSet(const std::vector<KeyValue>& kvs);
   Status MSetnx(const std::vector<KeyValue>& kvs, int32_t* ret);
   Status Set(const Slice& key, const Slice& value);
-  Status Setxx(const Slice& key, const Slice& value, int32_t* ret, int32_t ttl = 0);
+  Status Setxx(const Slice& key, const Slice& value, int32_t* ret, uint64_t ttl = 0);
   Status SetBit(const Slice& key, int64_t offset, int32_t value, int32_t* ret);
-  Status Setex(const Slice& key, const Slice& value, int32_t ttl);
-  Status Setnx(const Slice& key, const Slice& value, int32_t* ret, int32_t ttl = 0);
-  Status Setvx(const Slice& key, const Slice& value, const Slice& new_value, int32_t* ret, int32_t ttl = 0);
+  Status Setex(const Slice& key, const Slice& value, uint64_t ttl);
+  Status Setnx(const Slice& key, const Slice& value, int32_t* ret, uint64_t ttl = 0);
+  Status Setvx(const Slice& key, const Slice& value, const Slice& new_value, int32_t* ret, uint64_t ttl = 0);
   Status Delvx(const Slice& key, const Slice& value, int32_t* ret);
   Status Setrange(const Slice& key, int64_t start_offset, const Slice& value, int32_t* ret);
   Status Strlen(const Slice& key, int32_t* len);
